isolate problematic tests
